+++ contract/manager_test.go
@@ -618,7 +618,6 @@ func TestManagerWithCustomHandlers(t *testing.T) {
 			expectedTypes := []types.ContractType{
 				types.ContractTypeBoarding,
 				types.ContractTypeDefault,
-				types.ContractTypeHTLC,
 				types.ContractTypeVHTLC,
 			}
 			got := mgr.Registry().SupportedTypes()
@@ -633,7 +632,6 @@ func TestManagerWithCustomHandlers(t *testing.T) {
 				types.ContractTypeBoarding,
 				types.ContractType("custom"),
 				types.ContractTypeDefault,
-				types.ContractTypeHTLC,
 				types.ContractTypeVHTLC,
 			}
 			got := mgr.Registry().SupportedTypes()
