+++ identity/doc.go
@@ -66,10 +66,9 @@
 //
 // [Identity.SignMessage] schnorr-signs a raw message hash with the
 // account's first key. The concrete HD identity also implements
-// [KeyedTaprootSigner] for BTC/taproot flows that need to sign with a
-// specific wallet key id. [Identity.NewVtxoTreeSigner] returns a
-// short-lived signer session compatible with arkd's VTXO tree musig2
-// flow.
+// [KeyedPreimageSigner] for deterministic swap preimage derivation.
+// [Identity.NewVtxoTreeSigner] returns a short-lived signer session
+// compatible with arkd's VTXO tree musig2 flow.
 //
 // # Concurrency
 //
