+++ swap/chainswap_monitor.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"time"
 
-	arkidentity "github.com/arkade-os/arkd/pkg/client-lib/identity"
 	"github.com/arkade-os/go-sdk/swap/boltz"
 	"github.com/arkade-os/go-sdk/vhtlc"
+	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/chaincfg"
 	log "github.com/sirupsen/logrus"
 )
@@ -23,7 +23,7 @@ func (h *SwapHandler) monitorAndClaimArkToBtcSwap(
 	network *chaincfg.Params,
 	eventCallback ChainSwapEventCallback,
 	unilateralRefundCallback func(swapId string, opts vhtlc.Opts) error,
-	btcClaimKeyRef *arkidentity.KeyRef,
+	btcClaimKey *btcec.PrivateKey,
 	preimage []byte,
 	btcDestinationAddress string,
 	swapResp *boltz.CreateChainSwapResponse,
@@ -52,7 +52,7 @@ func (h *SwapHandler) monitorAndClaimArkToBtcSwap(
 			Swap:                     swap,
 		},
 		network,
-		btcClaimKeyRef,
+		btcClaimKey,
 		preimage,
 		btcDestinationAddress,
 		swapResp,
@@ -67,7 +67,7 @@ func (h *SwapHandler) monitorBtcToArkChainSwap(
 	ctx context.Context,
 	eventCallback ChainSwapEventCallback,
 	preimage []byte,
-	refundKeyRef *arkidentity.KeyRef,
+	refundKey *btcec.PrivateKey,
 	swapResp *boltz.CreateChainSwapResponse,
 	swap *ChainSwap,
 ) {
@@ -80,7 +80,7 @@ func (h *SwapHandler) monitorBtcToArkChainSwap(
 			Swap:          swap,
 		},
 		preimage,
-		refundKeyRef,
+		refundKey,
 		swapResp,
 	)
